Refactor channel-delete view to be responsive
